refactor: create a picking handler class handles picking for CAD models
