+++ viewer/packages/cad-model/index.ts
@@ -5,9 +5,9 @@
 export { CadManager } from './src/CadManager';
 export { GeometryFilter } from './src/GeometryFilter';
 export { CadModelSectorLoadStatistics } from './src/CadModelSectorLoadStatistics';
-export { intersectCadNodes } from './src/picking';
 export { createCadManager } from './src/createCadManager';
 export { BoundingBoxClipper } from './src/BoundingBoxClipper';
 export { InternalRevealCadOptions } from './src/InternalRevealCadOptions';
 export { Cognite3DModel } from './src/Cognite3DModel';
 export { WellKnownUnit } from './src/WellKnownUnit';
+export { PickingHandler } from './src/PickingHandler';
