+++ viewer/packages/cad-model/src/PickingHandler.test.ts
@@ -7,10 +7,12 @@ import * as THREE from 'three';
 import { CadNode } from '@reveal/rendering';
 import { IntersectInput } from '@reveal/model-base';
 
-import { intersectCadNodes } from './picking';
 import { createGlContext } from '../../../test-utilities';
+import { PickingHandler } from './PickingHandler';
 
-describe('intersectCadNodes', () => {
+describe(PickingHandler.name, () => {
+  let pickingHandler: PickingHandler;
+
   const camera = new THREE.PerspectiveCamera();
 
   const context = createGlContext(64, 64, { preserveDrawingBuffer: true });
@@ -28,13 +30,17 @@ describe('intersectCadNodes', () => {
   };
   const cadNode: CadNode = new THREE.Object3D() as any;
 
+  beforeEach(() => {
+    pickingHandler = new PickingHandler();
+  });
+
   test('no nodes, returns empty array', () => {
-    const intersections = intersectCadNodes([], input);
+    const intersections = pickingHandler.intersectCadNodes([], input);
     expect(intersections).toBeEmpty();
   });
 
   test('single node that does not intersect, returns empty array', () => {
-    const intersections = intersectCadNodes([cadNode], input);
+    const intersections = pickingHandler.intersectCadNodes([cadNode], input);
     expect(intersections).toBeEmpty();
   });
 });
